LayoutTests/imported/w3c/web-platform-tests/domxpath/fn-translate-expected.txt

^^^bar$$$^^^barfoo

FAIL translate() arguments depending on the context node assert_equals: expected "^^^foo$$$" but got "bar$$$"
